bug fix - skip only asyncio thread writes
The original solution was not working, bug was not detected as we ignored all error fired by execution on the preview thread.

But - it is causing real issues when IPython prints from prompt_toolkit async calls, for example when TerminalInteractiveShell.editing_mode=vi and IPython is changing states from insert to visual.